Crate jetscii

A tiny library to efficiently search strings for ASCII characters.

Example

use jetscii::AsciiChars;
let mut search = AsciiChars::new();
search.push(b'-');
search.push(b':');
let part_number = "86-J52:rev1";
let parts: Vec<_> = part_number.split(search.with_fallback(|c| {
    c == b'-' || c == b':'
})).collect();
assert_eq!(&parts, &["86", "J52", "rev1"]);

For maximum performance, you can create the searcher as a constant item. Print an existing AsciiChars with the debug formatter to get the appropriate invocation.

use jetscii::AsciiChars;
let search = AsciiChars { needle: 0x0000000000002d3a, count: 2 };
let part_number = "86-J52:rev1";
let parts: Vec<_> = part_number.split(search.with_fallback(|c| {
    c == b'-' || c == b':'
})).collect();
assert_eq!(&parts, &["86", "J52", "rev1"]);

Structs

AsciiChars
Searches a string for a set of ASCII characters. Up to 8 characters may be used.

AsciiCharsSearcher
An implementation of Searcher using AsciiChars

AsciiCharsWithFallback
Provides a hook for a user-supplied fallback implementation, used when the optimized instructions are not available.