Struct jetscii::AsciiChars

pub struct AsciiChars {
    pub needle: u64,
    pub count: u8,
}

Searches a string for a set of ASCII characters. Up to 8 characters may be used.

The instance variables are public to allow creating a AsciiChars as a constant item. This is temporary until Rust has better compile-time function evaluation, so consider this an unstable interface.

Fields

needle

count

Methods

impl AsciiChars

fn new() -> AsciiChars

fn push(&mut self, byte: u8)

Add a new ASCII character to the set to search for.

Panics

If you add more than 8 characters.

If you add a non-ASCII byte.

fn with_fallback<F>(self, fallback: F) -> AsciiCharsWithFallback<F> where F: Fn(u8) -> bool

Builds a searcher with a fallback implementation for when the optimized version is not available. The fallback should search for the exact same set of characters.